PM / Domains: Fix asynchronous execution of *noirq() callbacks
As the PM core may invoke the *noirq() callbacks asynchronously,
the current lock-less approach in genpd doesn't work. The consequence
is that we may find concurrent operations racing to power on/off the
PM domain.

As of now, no immediate errors have been reported, but it's probably
only a matter time. Therefor let's fix the problem now before this
becomes a real issue, by deploying the locking scheme to the relevant
functions.

Reported-by: Brian Norris <briannorris@chromium.org>
Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>