platform/x86: thinkpad_acpi: Add support for status LED
This allows the control of the red status LED, which is the dot of the "i"
in the word "ThinkPad" on the outside cover of newer models.

In the manual, both this LED and the power LED are referred to as
the "system-status indicators" without distinction between the two, so
I chose "status" as the LED name.

Signed-off-by: Adam Goode <agoode@google.com>
[andy: massaged title of the patch]
Signed-off-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>